Code/source/requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
opencv-python>=4.8.0
torch>=2.0.0
albumentations>=1.3.0
numpy>=1.24.0
PyYAML>=6.0
scikit-learn>=1.3.0
gradio>=4.0.0
